Added methods for searching movies by genre, filters and title. Distribution of static and media content has been implemented

+++ src/main/java/com/ivanhorlov/moviesbackend/dtos/MovieListResponse.java
@@ -11,6 +11,6 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MovieListResponse {
-    private int total_pages;
+    private long total_pages;
     private List<Movie> movie_list;
 }
